Update newline_placeholder
In some cases, `compare_answer` could confuse parts of your typed-in
text, with sequences of characters found within the string
"typeboxnewline", especially when your input text varies a lot from
the expected answer, leading to the text "__typeboxnewline__" sometimes
appearing in the UI, and your input answer has multiple newlines.

Updated the newline_placeholder variable to now consist entirely of
unicode characters that look similar to the original value, but which
are entirely unicode values and are very unlikely to match regular
user input.